Backend/stock_service/requirements.txt
sqlalchemy
pydantic
fastapi
uvicorn
pymysql
yfinance
pandas
python-jose[cryptography]
redis
